Crate tunneler_core

This crate provides a simply way to start a tunneler server and client that can also easily be integradted into other projects allowing you to expose your services running in a private network to be exposed through a public server

Modules

client
Provides all the Client related functionality

message
Messages are used for all Communication between Server and Client

metrics
All the Metrics related functionality

streams
Provides all the Stream/Queue related functionality

Structs

Destination
Describes a simple external Address consisting of an IP and Port

Details
The Details about a single Connection

Enums

DetailsIP
An IPV4 or IPV6 Address