Trait tunneler_core::client::Sender

pub trait Sender {
    type SendingError: Debug;
    #[must_use]
    fn send_msg<'life0, 'async_trait>(
        &'life0 self, 
        data: Vec<u8>, 
        length: u64
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::SendingError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

A Generic trait that abstract away the actual underlying type and implementation for sending Data back to the User over that connection

Associated Types

type SendingError: Debug

The Error-Type that the Sender will return if it encounters any problems

Required methods

#[must_use]
fn send_msg<'life0, 'async_trait>(
&'life0 self,
data: Vec<u8>,
length: u64
) -> Pin<Box<dyn Future<Output = Result<(), Self::SendingError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Sends a single Message over the Connection