Struct tunneler_core::client::Client

pub struct Client<M> { /* fields omitted */ }

The Client instance in general that is responsible for handling all the interactions with the Server

Implementations

impl Client<Empty>

pub fn new(server: Destination, external_port: u16, key: Vec<u8>) -> Self

Creates a new Client instance that is configured to connect to the given Server Destination and authenticate using the provided Key.

This uses an empty Metrics-Collector meaning that no metrics will be collected

impl<M> Client<M> where
M: Metrics + Send + Sync + 'static,

pub fn new_metrics(
server: Destination,
external_port: u16,
key: Vec<u8>,
metrics_collector: M
) -> Self

Behaves the same as the new implementation with the difference being that you can specify and provide your own Metrics-Collector.

pub async fn start<H>(self, handler: Arc<H>) -> ! where
H: Handler + Send + Sync + 'static,

This starts the client and all the needed tasks

This function essentially never returns and should therefor be run in an independant task

The Handler will be passed as arguments:

The ID of the new user-connection

A Reader where all the Messages for this user can be read from

A Writer which can be used to send data back to the user

The handler_data that can be used to share certain information when needed

The start_handler is only ever called once for every new connection in a seperate tokio::Task All Messages the Handler receives only Data or EOF Messages